feat: move new declaration next to new simulateur
