+++ packages/app/src/app/(default)/index-egapro_/declaration/declarationFunnelConfiguration.ts
