+++ packages/app/src/app/(default)/index-egapro_/declaration/periode-reference/page.tsx
@@ -1,0 +1,17 @@
+import { type FunnelKey } from "../declarationFunnelConfiguration";
+import { DeclarationStepper } from "../DeclarationStepper";
+import { PeriodeReferenceForm } from "./PeriodeReferenceForm";
+
+const stepName: FunnelKey = "periode-reference";
+
+const InformationsEntreprisePage = () => {
+  return (
+    <>
+      <DeclarationStepper stepName={stepName} />
+
+      <PeriodeReferenceForm />
+    </>
+  );
+};
+
+export default InformationsEntreprisePage;
