+++ packages/app/src/app/(default)/index-egapro_/declaration/remunerations/page.tsx
@@ -1,0 +1,17 @@
+import { type FunnelKey } from "../declarationFunnelConfiguration";
+import { DeclarationStepper } from "../DeclarationStepper";
+import { RemunerationForm } from "./RemunerationForm";
+
+const stepName: FunnelKey = "remunerations";
+
+const RemunerationPage = () => {
+  return (
+    <>
+      <DeclarationStepper stepName={stepName} />
+
+      <RemunerationForm />
+    </>
+  );
+};
+
+export default RemunerationPage;
