--- packages/app/src/app/index-egapro_/declaration/remunerations-resultat/page.tsx
@@ -1,28 +1,0 @@
-"use client";
-
-import { Stepper } from "@codegouvfr/react-dsfr/Stepper";
-import { ClientOnly } from "@components/utils/ClientOnly";
-import { useDeclarationFormManager } from "@services/apiClient/useDeclarationFormManager";
-
-import { nbSteps } from "../../declarationFunnelConfiguration";
-import { RemunerationCSPResultatForm } from "./RemunerationCSPResultatForm";
-
-const title = "Résultat final de l’écart de rémunération entre les femmes et les hommes";
-
-const RemunerationResultatPage = () => {
-  const { formData } = useDeclarationFormManager();
-
-  const nextTitle = `Écart de taux d'augmentations individuelles ${
-    formData.entreprise?.tranche === "50:250" && "(hors promotion)"
-  } entre les femmes et les hommes`;
-
-  return (
-    <ClientOnly>
-      <Stepper currentStep={6} nextTitle={nextTitle} stepCount={nbSteps} title={title} />
-
-      <RemunerationCSPResultatForm />
-    </ClientOnly>
-  );
-};
-
-export default RemunerationResultatPage;
