--- packages/app/src/app/index-egapro_/declaration/remunerations/page.tsx
@@ -1,17 +1,0 @@
-import { type FunnelKey } from "../../declarationFunnelConfiguration";
-import { DeclarationStepper } from "../DeclarationStepper";
-import { RemunerationForm } from "./RemunerationForm";
-
-const stepName: FunnelKey = "remunerations";
-
-const RemunerationPage = () => {
-  return (
-    <>
-      <DeclarationStepper stepName={stepName} />
-
-      <RemunerationForm />
-    </>
-  );
-};
-
-export default RemunerationPage;
