--- packages/app/src/app/index-egapro_/declaration/ues/page.tsx
@@ -1,17 +1,0 @@
-import { type FunnelKey } from "../../declarationFunnelConfiguration";
-import { DeclarationStepper } from "../DeclarationStepper";
-import { UESForm } from "./UESForm";
-
-const stepName: FunnelKey = "ues";
-
-const InformationsUESPage = () => {
-  return (
-    <>
-      <DeclarationStepper stepName={stepName} />
-
-      <UESForm />
-    </>
-  );
-};
-
-export default InformationsUESPage;
