--- packages/app/src/components/next13/README.md
@@ -1,3 +1,0 @@
-## README
-
-Ajouter ici les composants migrés en React DSFR @codegouvfr/react-dsfr
