--- packages/app/src/components/next13/RecapSection.module.css
@@ -1,9 +1,0 @@
-.list {
-    list-style: none;
-    padding: 0;
-}
-
-
-.list > li + li {
-    margin-top: 0.375rem;
-}
